Import error flagged by code-climate
